refactor: Criado diretório para agrupar nodes
